Added new HuskyContext for all relevant husky steps
